backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
boto3==1.35.24
pypdf==5.0.1
python-docx==1.1.2
pdfplumber==0.11.4
